Move rf-input logic to mixin
