+++ lib/mixins/index.js
@@ -1,1 +1,2 @@
 import './rf-input-helpers'
+import './rf-base-input'
